+++ CHANGELOG.md
@@ -11,14 +11,27 @@
 
 
 
+## "version": "2.1.4"
+## "date": "2022.05.02"
+## "author": "cdl"
+
+- 🎯 优化 图标选择器参数，设置默认 prepend 和 placeholder，新增弹窗标题
+- 🚧 调整图标选择器 Popover 气泡卡片触发方式改为虚拟触发
+- 🎉 新增 图标选择器新增 title 标题参数
+
+
+
 ## "version": "2.1.3"
 ## "date": "2022.05.01"
+## "author": "cdl"
+
 - 🐞 修复 图标选择器展示页面引入的图标选择器组件文件名大小写敏感问题
 
 
 
 ## "version": "2.1.2"
 ## "date": "2022.05.01"
+## "author": "cdl"
 
 - 🎯 优化 图标选择器在不同分辨率下的展示方式
 - 🎉 新增 图标选择器展示页面 前往体验：[滚动通知栏](https://vue3-vite2-admin.vercel.app/#/func/selector)
@@ -28,6 +41,7 @@
 
 ## "version": "2.1.1"
 ## "date": "2022.05.01"
+## "author": "cdl"
 
 - 🐞 修复 深色模式下，刷新浏览器之后深色模式丢失
 - 🎯 优化 css 变量
@@ -37,6 +51,7 @@
 
 ## "version": "2.1.0"
 ## "date": "2022.04.29"
+## "author": "cdl"
 
 - 🎉 新增 tagsView 支持自定义 tagsView 名称（文章详情时有用），前往体验：[路由参数/普通路由](https://lyt-top.gitee.io/vue-next-admin-preview/#/params/common)
 - 🎉 新增 普通路由和动态路由页面